Fix multiple writes to branch ref key when using source_root.
When using source_root paths outside that source_root would be processed
and due to errors in the processing loop logic would trigger the
callback that would ultimately trigger multiple writes to the branch ref
key in consul.  It also prematurely releases the lock on the branch.

This happens because the logic for skipping paths outside the
source_path calls check_pending which decrements the pending_record
count.  When the record count is at or below zero then it would call
the upstream callback and trigger the ref key write and the removal
of the locks.

Additionally, the callback was wrapped in _.once() implying an attempt
to prevent the callback from being called multiple times.  But this
fails because the code called the cb() function and passed the return
to _.once, instead of using _.once to create a new variable.

In the handling for the common_properties file, errors in processing
the consul updates were prefixed with "Some consul updates failed:",
but these errors are appended to the existing errors which would then
be prefixed by the same string further up, resulting in the prefix
appearing multiple times in the logs.

Fix removes the pending_records counter entirely and uses _.after()
to wrap the callback so that the callback is only called after all
records are processed.  This means that it is important that calling
check_pending() must be called on every record exactly once. The old
logic technically would allow you to call it multiple times as long
as you incremented the same number of times.

Finally, move the error handler for an unknown git status to also
call check_pending and provide the record.path for easier debugging.

While we're at it make it obvious which records are skipped due to
being outside the source_root when doing trace logging.